created common function
